fixup: remove node specific stuff
