+++ lib/api/api-request.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const Readable = require('../node/readable')
+const Readable = require('./readable')
 const {
   InvalidArgumentError,
   RequestAbortedError
@@ -8,27 +8,6 @@ const {
 const util = require('../core/util')
 const { AsyncResource } = require('async_hooks')
 const { addSignal, removeSignal } = require('./abort-signal')
-
-const kAbort = Symbol('abort')
-
-class RequestResponse extends Readable {
-  constructor (resume, abort) {
-    super({ autoDestroy: true, read: resume })
-    this[kAbort] = abort
-  }
-
-  _destroy (err, callback) {
-    if (!err && !this._readableState.endEmitted) {
-      err = new RequestAbortedError()
-    }
-
-    if (err) {
-      this[kAbort]()
-    }
-
-    callback(err)
-  }
-}
 
 class RequestHandler extends AsyncResource {
   constructor (opts, callback) {
@@ -92,7 +71,7 @@ class RequestHandler extends AsyncResource {
       return
     }
 
-    const body = new RequestResponse(resume, abort)
+    const body = new Readable(resume, abort)
 
     this.callback = null
     this.res = body
