+++ lib/api/readable.js
@@ -6,14 +6,14 @@
 
 const assert = require('assert')
 const { Readable } = require('stream')
-const { InvalidArgumentError, RequestAbortedError } = require('../core/errors')
+const { RequestAbortedError } = require('../core/errors')
 
-let StringDecoder
 let Blob
 
 const kConsume = Symbol('kConsume')
 const kReading = Symbol('kReading')
 const kBody = Symbol('kBody')
+const kAbort = Symbol('abort')
 
 const kTextType = 1
 const kBlobType = 2
@@ -21,11 +21,12 @@ const kArrayBufferType = 3
 const kJSONType = 4
 
 module.exports = class BodyReadable extends Readable {
-  constructor (opts) {
-    super(opts)
+  constructor (resume, abort) {
+    super({ autoDestroy: true, read: resume })
 
     this._readableState.dataEmitted = false
 
+    this[kAbort] = abort
     this[kConsume] = null
     this[kBody] = null
 
@@ -41,6 +42,15 @@ module.exports = class BodyReadable extends Readable {
       // Node < 16
       return this
     }
+
+    if (!err && !this._readableState.endEmitted) {
+      err = new RequestAbortedError()
+    }
+
+    if (err) {
+      this[kAbort]()
+    }
+
     return super.destroy(err)
   }
 
@@ -81,20 +91,13 @@ module.exports = class BodyReadable extends Readable {
     return this.off(ev, ...args)
   }
 
-  push (chunk, encoding) {
+  push (chunk) {
     if (this[kConsume] && chunk !== null && !this[kReading]) {
-      // Fast path.
-      return consumePush(
-        this[kConsume],
-        chunk,
-        encoding || this._readableState.defaultEncoding
-      )
+      consumePush(this[kConsume], chunk)
+      return true
+    } else {
+      return super.push(chunk)
     }
-
-    const pushed = super.push(chunk, encoding)
-    const consumed = consumePush(this[kConsume], chunk, encoding)
-
-    return pushed && consumed
   }
 
   // https://fetch.spec.whatwg.org/#dom-body-text
@@ -138,7 +141,7 @@ module.exports = class BodyReadable extends Readable {
 
 // https://streams.spec.whatwg.org/#readablestream-locked
 function isLocked (self) {
-  // consume is implicit lock
+  // Consume is an implicit lock.
   return (self[kBody] && self[kBody].locked === true) || self[kConsume]
 }
 
@@ -173,11 +176,7 @@ async function consume (stream, type) {
       resolve,
       reject,
       length: 0,
-      decoder: undefined,
-      body: undefined,
-      reading: false,
-      pushed: false,
-      ended: false
+      body: undefined
     }
 
     stream
@@ -251,61 +250,17 @@ function consumeEnd (consume) {
 }
 
 function consumePush (consume, chunk, encoding) {
-  if (!consume) {
-    return true
-  }
-
-  if (consume.ended) {
-    return false
-  }
-
-  if (chunk === null) {
-    consume.ended = true
-    consume.stream.read()
-    return false
-  }
-
-  consume.pushed = true
-
   if (consume.type === kTextType || consume.type === kJSONType) {
     consumePushString(consume, chunk, encoding)
   } else {
     consumePushBuffer(consume, chunk, encoding)
   }
-
-  if (!consume.stream[kReading] && !consume.reading) {
-    consume.reading = true
-    process.nextTick(consumeReadMore, consume)
-  }
-
-  return true
 }
 
 function consumePushString (consume, chunk, encoding) {
   if (!consume.body) {
     consume.body = ''
   }
-
-  if (typeof chunk === 'string') {
-    if (consume.decoder) {
-      chunk = consume.decoder.write(Buffer.from(chunk, encoding))
-    } else if (encoding !== 'utf8') {
-      chunk = Buffer.from(chunk, encoding).toString()
-    }
-  } else if (ArrayBuffer.isView(chunk)) {
-    if (!consume.decoder) {
-      if (!StringDecoder) {
-        StringDecoder = require('string_decoder').StringDecoder
-      }
-      consume.decoder = new StringDecoder('utf8')
-    }
-    chunk = consume.decoder.write(chunk)
-  } else {
-    // TODO: This does not support objectMode.
-    // TODO: Use Node error?
-    throw new InvalidArgumentError('chunk')
-  }
-
   consume.length += chunk.length
   consume.body += chunk
 }
@@ -314,31 +269,8 @@ function consumePushBuffer (consume, chunk, encoding) {
   if (!consume.body) {
     consume.body = []
   }
-
-  if (typeof chunk === 'string') {
-    chunk = Buffer.from(chunk, encoding)
-  } else if (!ArrayBuffer.isView(chunk)) {
-    // TODO: This does not support objectMode.
-    // TODO: Use Node error?
-    throw new InvalidArgumentError('chunk')
-  }
-
   consume.length += chunk.byteLength
   consume.body.push(chunk)
-}
-
-function consumeReadMore (consume) {
-  consume.pushed = true
-
-  while (consume.pushed && !consume.stream[kReading]) {
-    consume.pushed = false
-    // TODO: This does not support thenable return.
-    // TODO: This does not support throwing in _read().
-    consume.stream._read()
-  }
-
-  consume.pushed = false
-  consume.reading = false
 }
 
 function consumeFinish (consume, err) {
